+++ Mokkivaraus/Mokkivaraus.csproj
@@ -60,19 +60,19 @@
     <Compile Include="Aloitussivu.Designer.cs">
       <DependentUpon>Aloitussivu.cs</DependentUpon>
     </Compile>
-    <Compile Include="Toimipisteet.cs">
+    <Compile Include="Toimintaalueet.cs">
       <SubType>Form</SubType>
     </Compile>
-    <Compile Include="Toimipisteet.Designer.cs">
-      <DependentUpon>Toimipisteet.cs</DependentUpon>
+    <Compile Include="Toimintaalueet.Designer.cs">
+      <DependentUpon>Toimintaalueet.cs</DependentUpon>
     </Compile>
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <EmbeddedResource Include="Aloitussivu.resx">
       <DependentUpon>Aloitussivu.cs</DependentUpon>
     </EmbeddedResource>
-    <EmbeddedResource Include="Toimipisteet.resx">
-      <DependentUpon>Toimipisteet.cs</DependentUpon>
+    <EmbeddedResource Include="Toimintaalueet.resx">
+      <DependentUpon>Toimintaalueet.cs</DependentUpon>
     </EmbeddedResource>
     <EmbeddedResource Include="Properties\Resources.resx">
       <Generator>ResXFileCodeGenerator</Generator>
